// src/app/features/orders-page/enums/orders.paths.enum.ts
export enum ORDERS_API_PATHS {
     LIST = 'orders',
     //  LIST_STATUS_ID = 'orders/status/',
     LAST = 'orders/last',
     CLIENT_LAST = 'orders/last/client/',
     IN_PRODUCTION = 'orders/inproduction',
     TO_TAKE = 'orders/totake',
     UNSENT = 'orders/unsent',

     CLIENT_ORDERS = 'orders/client/',
     DETAILS = 'myorder/',
     DETAILS_SHEET = 'myorder/sheet/',
     ORDER_DETAILS = 'orders/show/',
     ORDER_NEW = 'orders/new/',
     ORDER_ADD = 'orders/add/',

     PAYMENT = 'orders/payment/',
     PAY_IN_ADVANCE = 'orders/advance-pay/',

     STATUS_LIST = 'orders/status/list',
     LIST_STATUS_ID = 'orders/status/',
     STATUS_CHANGE = 'orders/status/change/',

     UPDATE = 'orders/update/',
}